+++ importd.py
@@ -40,7 +40,6 @@ class D(object):
                 attributes = [attributes]
             callback(module_name, attributes)             # check if its a decorated view from importd
 
-
     def _import_django(self):
         def set_attr(module_name, attributes):
             import importlib
@@ -49,9 +48,12 @@ class D(object):
                 setattr(self, attribute, getattr(module, attribute))
         self._iterate_imports(set_attr)
 
-        self.wsgi_application = self.get_wsgi_application()
-
-
+        try:
+            from django.core.wsgi import get_wsgi_application
+            self.wsgi_application = self.get_wsgi_application()
+        except ImportError:
+            import django.core.handlers.wsgi
+            self.wsgi_application = django.core.handlers.wsgi.WSGIHandler()
 
     def dotslash(self, pth):
         import os
@@ -132,13 +134,16 @@ class D(object):
         def process_view(self, request, view_func, view_args, view_kwargs):
             from django.shortcuts import render_to_response
             from django.template import RequestContext
-            from django.http import HttpResponse
+            try:
+                from django.http.response import HttpResponseBase as RBase
+            except ImportError:
+                from django.http import HttpResponse as RBase
 
             from fhurl import JSONResponse
             res = view_func(request, *view_args, **view_kwargs)
             if isinstance(res, basestring):
                 res = res, {}
-            if isinstance(res, HttpResponse): return res
+            if isinstance(res, RBase): return res
             if isinstance(res, tuple):
                 template_name, context = res
                 res = render_to_response(