+++ tests/basic_tests/app/signals.py
@@ -1,0 +1,3 @@
+from django.conf import settings
+
+settings.SIGNALS_IMPORTED=True